+++ CHANGELOG.md
@@ -24,6 +24,7 @@ All notable changes to the Cosmy Wasmy extension will be documented in this file
 
 - For testnet and mainnet chain, all contracts by the selected account can be imported in a single click
 - Add option to simulate a contract execution in the tx view in sidebar
+- Added a command to query tx details from the tx hash 
 
 ### Changed